python3 - ch04_basics - task2 - redundant code removed

+++ ch04_basics.py
@@ -14,7 +14,7 @@ print(get_day_week(7, 3))
 # The basics: Practical Task 2
 def count_digits(n):
     if isinstance(n, int) and n > 0:
-        return int(math.ceil(math.log10(n + .1)))
+        return math.ceil(math.log10(n + .1))
     return 0
